[Refactor] API 요청 함수를 분리하고, 적절한 요청 메시지로 변경
