+++ frontend/src/api/index.ts
@@ -1,3 +1,4 @@
 export * from './capacity';
 export * from './getDirectory';
 export * from './getTrashFiles';
+export * from './file';
